+++ src/features/components/templates/trading-platform/app-banner/index.tsx
