+++ src/features/components/templates/trading-platform/hero/hero.module.scss
