+++ src/features/components/templates/trading-platform/hero/index.tsx
