+++ src/features/components/templates/trading-platform/how-to-start/index.tsx
