+++ src/features/components/templates/trading-platform/index.ts
