+++ src/features/components/templates/trading-platform/numbers/index.tsx
