+++ src/features/components/templates/trading-platform/numbers/numbers.module.scss
