+++ src/features/components/templates/trading-platform/other-platforms/index.tsx
