+++ src/features/components/templates/trading-platform/what-have/what-have.module.scss
