+++ src/features/components/templates/trading-platform/why-trade/index.tsx
