+++ src/features/pages/ctrader/index.tsx
@@ -1,13 +1,11 @@
 import React from 'react'
 import { ctrader_data } from './data'
 import {
-    TradingPlatformHero,
     TradingPlatformHowToStart,
     TradingPlatformNumbers,
-    TradingPlatformOther,
     TradingPlatformWhatHave,
     TradingPlatformWhyTrade,
-} from 'features/components/templates/trade-platform'
+} from 'features/components/templates/trading-platform'
 import MainNav from 'features/components/templates/navigation/main-nav'
 import Footer from 'features/components/templates/footer/'
 import Layout from 'features/components/templates/layout'
@@ -32,12 +30,14 @@ const CTrader = () => {
                     '_t_cTrader is our intuitive CFD trading platform where you can trade forex, commodities, stocks & indices, cryptocurrencies, and derived indices._t_',
                 )}
             />
-            <TradingPlatformHero trading_platform_hero={trading_platform_hero} />
+            {/*need provide new*/}
+            {/*<TradingPlatformHero trading_platform_hero={trading_platform_hero} />*/}
             <TradingPlatformNumbers trading_platform_numbers={trading_platform_numbers} />
             <TradingPlatformWhyTrade trading_platform_why_trade={trading_platform_why_trade} />
             <TradingPlatformWhatHave trading_platform_what_have={trading_platform_what_have} />
             <TradingPlatformHowToStart trading_platform_start={trading_platform_start} />
-            <TradingPlatformOther trading_platform_other={trading_platform_other} />
+            {/*need to complete*/}
+            {/*<TradingPlatformOther trading_platform_other={trading_platform_other} />*/}
             <Footer />
         </Layout>
     )
